+++ src/main/java/com/phantomwing/rusticdelight/tags/CompatibilityTags.java
@@ -47,6 +47,10 @@ public class CompatibilityTags
     // Farmer's Respite
     public static final String FARMERS_RESPITE = "farmersrespite";
 
+    // Supplementaries
+    public static final String SUPPLEMENTARIES = "supplementaries";
+    public static final TagKey<Item> COOKIES = externalItemTag(SUPPLEMENTARIES, "cookies");
+
     private static TagKey<Item> externalItemTag(String modId, String path) {
         return ItemTags.create(new ResourceLocation(modId, path));
     }