+++ changelog.md
@@ -2,11 +2,23 @@
 ### Additions
 - Added Bag of Roasted Coffee Beans
 - Added Japanese translations (thank you EndilCrafter!)
-
-### Changes
-- Added compatibility with Farmer's Respite
+- Added compatibility with Botany Pots:
+  - Cotton, Coffee and Bell Peppers can now be grown in a Botany Pot
+  - Wild Cotton, Wild Coffee and Wild Bell Peppers are also supported
+- Added compatibility with Farmer's Respite:
   - Brew Farmer's Respite Coffee with Rustic Delight's Roasted Coffee Beans using a Kettle
   - Rustic Delight Coffee (and variants) can be brewed with Farmer's Respite Coffee Beans using the Cooking Pot
+  - Rustic Delight Coffee variants can be crafted using Farmer's Respite Coffee
+- Added compatibility with Supplementaries:
+  - Cherry Blossom Cookie can now be put in a jar
+  
+### Changes
+- Golden Coffee Beans can now also be crafted with Roasted Coffee Beans
+- Roasted Coffee Beans can now be crafted into Brown Dye
+- Made the following items compostable
+  - Cherry Blossom Cookie
+  - Cherry Blossom Cheesecake
+  - Slice of Cherry Blossom Cheesecake
 - Updated textures for Coffee Beans, Roasted Coffee Beans, Coffee Beans Bag
 
 # 1.3.0